+++ test/unit/fuzz-tools.test.js
@@ -128,6 +128,7 @@ import * as nephroV127 from '../../lib/nephro-v127.js';
 import * as renalV128 from '../../lib/renal-v128.js';
 import * as acidbaseV129 from '../../lib/acidbase-v129.js';
 import * as uroV130 from '../../lib/uro-v130.js';
+import * as emsV149 from '../../lib/ems-v149.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -200,6 +201,7 @@ const MODULES = {
   'renal-v128.js': renalV128,
   'acidbase-v129.js': acidbaseV129,
   'uro-v130.js': uroV130,
+  'ems-v149.js': emsV149,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];